Added donut.jsfuck
I didnt leave enough space to explain what i did in the comments on donut.jsfuck, so ill say it here: I just removed the spaces and linebreaks from donut.js, out it through the converter at jsfuck.com, copied the text and filled a donut with it.
To run the code, just run node and give it the file as the argument, just like you'd do with donut.js too.

I also fixed some of the formatting in info.json since it was bugging me a little.