Added new functionality to donut_generator
It can now calculate a donut shape with a desired minimum character count.
It isnt perfect, but it takes away a little bit of guessing and will make the donuts look a lil more uniform (if the devs use this at least).

I also updated the error / usage message to reflect the new functionality.